requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.4
python-multipart>=0.0.18
sqlalchemy>=2.0.36
pydantic-settings>=2.6.0
argon2-cffi>=23.1.0
click>=8.1.7
